feat(profile) - Made ProfileAdd modal modaular and added simple area
